==> R/CardiovascularDiseaseRisk/1.1_FramChart.R <==
# Title: ADRC Framingham Chart based on WHO North America 
# Author: Diana Hobbs
# Date: April 2022

#Functions to set up environment
rm(list=ls()) #removes all variables from the current environment
options(scipen = 999) #forces R not to use scientific notation

#### Set Up ####
# Load packages 
pacman::p_load(tidyverse,jtools,naniar,lubridate)

###### Load data and select columns of interest
data <- read.csv("./Data/Rabin_Replication_Demographics.csv")%>%
  mutate(CVD_Risk = case_when(
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "<5%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ ">=30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ ">=30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ ">=30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS<120 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS<120 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS<120 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS<120 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI<20 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=20 & BMI<25 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=25 & BMI<30 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Male" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "<5%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "<5%",
  
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "<5%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "5-10%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "<5%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "<5%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "<5%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==0 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "<5%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "5-10%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=40 & AGE<45 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "<5%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "5-10%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "10-20%",
  AGE>=45 & AGE<50 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "5-10%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "5-10%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "10-20%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=50 & AGE<55 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "5-10%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "10-20%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=55 & AGE<60 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "10-20%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ "20-30%",
  AGE>=60 & AGE<65 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "10-20%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "10-20%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ "20-30%",
  
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=65 & AGE<70 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS<120 ~ "10-20%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS<120 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=120 & BPSYS<140 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=140 & BPSYS<160 ~ "20-30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=160 & BPSYS<180 ~ ">=30%",
  
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI<20 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=20 & BMI<25 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=25 & BMI<30 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=30 & BMI<35 & BPSYS>=180 ~ ">=30%",
  AGE>=70 & AGE<105 & SMOKE==1 & SEX=="Female" & BMI>=35 & BPSYS>=180 ~ ">=30%"))

#### Final Data Set ####
write.csv(data,"./Data/Rabin_Replication_CVD-RF.csv",row.names = FALSE)
