==> R/CardiovascularDiseaseRisk/1.3_MR-Normalize.R <==
# Title: ADRC Normalizing MR to sample 
# Cleaning and merging files  
# Author: Diana Hobbs
# Date: April 2022

# Load packages & data
pacman::p_load(tidyverse,jtools,naniar,lubridate,janitor)

mri <- read.csv("./Data/Rabin_Replication_MRI.csv")%>%
  mutate(SampleMeanICV = mean(ICV))
  

##### Normalize the mri volumes ####
#lmTotHipp = lm(mri$hippocampus ~ mri$ICV)
#mri$hippocampus <- mri$hippocampus - ((lmTotHipp[["coefficients"]][["mri$ICV"]]) * (mri$ICV - mri$SampleMeanICV))
lm1=lm(mri$bankssts~mri$ICV)
lm2=lm(mri$caudal.anterior.cingulate~mri$ICV)
lm3=lm(mri$caudal.middle.frontal~mri$ICV)
lm4=lm(mri$cuneus~mri$ICV)
lm5=lm(mri$entorhinal~mri$ICV)
lm6=lm(mri$frontal.pole~mri$ICV)
lm7=lm(mri$fusiform~mri$ICV)
lm8=lm(mri$inferior.parietal~mri$ICV)
lm9=lm(mri$inferior.temporal~mri$ICV)
lm10=lm(mri$insula~mri$ICV)
lm11=lm(mri$isthmus.cingulate~mri$ICV)
lm12=lm(mri$lateral.occipital~mri$ICV)
lm13=lm(mri$lateral.orbitofrontal~mri$ICV)
lm14=lm(mri$lingual~mri$ICV)
lm15=lm(mri$medial.orbitofrontal~mri$ICV)
lm16=lm(mri$middle.temporal~mri$ICV)
lm17=lm(mri$paracentral~mri$ICV)
lm18=lm(mri$parahippocampal~mri$ICV)
lm19=lm(mri$pars.opercularis~mri$ICV)
lm20=lm(mri$pars.orbitalis~mri$ICV)
lm21=lm(mri$pars.triangularis~mri$ICV)
lm22=lm(mri$pericalcarine~mri$ICV)
lm23=lm(mri$postcentral~mri$ICV)
lm24=lm(mri$posterior.cingulate~mri$ICV)
lm25=lm(mri$precentral~mri$ICV)
lm26=lm(mri$precuneus~mri$ICV)
lm27=lm(mri$rostral.anterior.cingulate~mri$ICV)
lm28=lm(mri$rostral.middle.frontal~mri$ICV)
lm29=lm(mri$superior.frontal~mri$ICV)
lm30=lm(mri$superior.parietal~mri$ICV)
lm31=lm(mri$superior.temporal~mri$ICV)
lm32=lm(mri$supramarginal~mri$ICV)
lm33=lm(mri$temporal.pole~mri$ICV)
lm34=lm(mri$transverse.temporal~mri$ICV)
lm37=lm(mri$amygdala~mri$ICV)
lm38=lm(mri$caudate~mri$ICV)
lm41=lm(mri$hippocampus~mri$ICV)
lm43=lm(mri$pallidum~mri$ICV)
lm44=lm(mri$putamen~mri$ICV)
lm45=lm(mri$thalamus.proper~mri$ICV)
lm46=lm(mri$ventral.DC~mri$ICV)

mri$bankssts<-mri$bankssts-((lm1[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$caudal.anterior.cingulate<-mri$caudal.anterior.cingulate-((lm2[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$caudal.middle.frontal<-mri$caudal.middle.frontal-((lm3[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$cuneus<-mri$cuneus-((lm4[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$entorhinal<-mri$entorhinal-((lm5[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$frontal.pole<-mri$frontal.pole-((lm6[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$fusiform<-mri$fusiform-((lm7[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$inferior.parietal<-mri$inferior.parietal-((lm8[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$inferior.temporal<-mri$inferior.temporal-((lm9[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$insula<-mri$insula-((lm10[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$isthmus.cingulate<-mri$isthmus.cingulate-((lm11[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$lateral.occipital<-mri$lateral.occipital-((lm12[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$lateral.orbitofrontal<-mri$lateral.orbitofrontal-((lm13[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$lingual<-mri$lingual-((lm14[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$medial.orbitofrontal<-mri$medial.orbitofrontal-((lm15[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$middle.temporal<-mri$middle.temporal-((lm16[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$paracentral<-mri$paracentral-((lm17[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$parahippocampal<-mri$parahippocampal-((lm18[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$pars.opercularis<-mri$pars.opercularis-((lm19[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$pars.orbitalis<-mri$pars.orbitalis-((lm20[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$pars.triangularis<-mri$pars.triangularis-((lm21[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$pericalcarine<-mri$pericalcarine-((lm22[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$postcentral<-mri$postcentral-((lm23[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$posterior.cingulate<-mri$posterior.cingulate-((lm24[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$precentral<-mri$precentral-((lm25[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$precuneus<-mri$precuneus-((lm26[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$rostral.anterior.cingulate<-mri$rostral.anterior.cingulate-((lm27[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$rostral.middle.frontal<-mri$rostral.middle.frontal-((lm28[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$superior.frontal<-mri$superior.frontal-((lm29[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$superior.parietal<-mri$superior.parietal-((lm30[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$superior.temporal<-mri$superior.temporal-((lm31[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$supramarginal<-mri$supramarginal-((lm32[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$temporal.pole<-mri$temporal.pole-((lm33[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$transverse.temporal<-mri$transverse.temporal-((lm34[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$amygdala<-mri$amygdala-((lm37[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$caudate<-mri$caudate-((lm38[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$hippocampus<-mri$hippocampus-((lm41[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$pallidum<-mri$pallidum-((lm43[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$putamen<-mri$putamen-((lm44[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$thalamus.proper<-mri$thalamus.proper-((lm45[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
mri$ventral.DC<-mri$ventral.DC-((lm46[["coefficients"]][["mri$ICV"]])*(mri$ICV-mri$SampleMeanICV))
rm(lm1,lm2,lm3,lm4,lm5,lm6,lm7,lm8,lm9,lm10,lm11,lm12,lm13,lm14,lm15,lm16,lm17,lm18,lm19,lm20,lm21,lm22,lm23,lm24,lm25,lm26,lm27,
   lm28,lm29,lm30,lm31,lm32,lm33,lm34,lm37,lm38,lm41,lm43,lm44,lm45,lm46)
write.csv(mri,"./Data/Rabin_Replication_MRI_Normalized.csv",row.names = FALSE)
